Initialized diktat some tests
### What's done:
* Fixes after merge
