+++ diktat-rules/src/test/kotlin/org/cqfn/diktat/ruleset/smoke/DiktatSmokeTest.kt
@@ -16,8 +16,9 @@ import org.junit.jupiter.api.Test
 
 // fixme: run as a separate maven goal/module?
 class DiktatSmokeTest : FixTestBase("test/smoke",
-        DiktatRuleSetProvider().get(),
-        { lintError, _ -> unfixedLintErrors.add(lintError) }
+        { DiktatRuleSetProvider() },
+        { lintError, _ -> unfixedLintErrors.add(lintError) },
+        null
 ) {
     companion object {
         private val unfixedLintErrors: MutableList<LintError> = mutableListOf()
@@ -35,10 +36,10 @@ class DiktatSmokeTest : FixTestBase("test/smoke",
         unfixedLintErrors.assertEquals(
                 LintError(1, 1, "$DIKTAT_RULE_SET_ID:file-naming", "${FILE_NAME_INCORRECT.warnText()} Example1Test.kt_copy", true), // todo this is a false one
                 LintError(1, 1, "$DIKTAT_RULE_SET_ID:file-naming", "${FILE_NAME_MATCH_CLASS.warnText()} Example1Test.kt_copy vs Example", true), // todo this is a false one
+                LintError(1, 1, "$DIKTAT_RULE_SET_ID:kdoc-comments", "${MISSING_KDOC_CLASS_ELEMENTS.warnText()} foo", false), // todo what's with offset?
                 LintError(1, 1, "$DIKTAT_RULE_SET_ID:kdoc-formatting", "${KDOC_NO_EMPTY_TAGS.warnText()} @return", false),
-                LintError(3, 17, "$DIKTAT_RULE_SET_ID:kdoc-comments", "${MISSING_KDOC_TOP_LEVEL.warnText()} Example", false),
-                LintError(3, 37, "$DIKTAT_RULE_SET_ID:kdoc-comments", "${MISSING_KDOC_CLASS_ELEMENTS.warnText()} isValid", false),
-                LintError(7, 3, "$DIKTAT_RULE_SET_ID:kdoc-comments", "${MISSING_KDOC_CLASS_ELEMENTS.warnText()} foo", false)
+                LintError(7, 14, "$DIKTAT_RULE_SET_ID:kdoc-comments", "${MISSING_KDOC_TOP_LEVEL.warnText()} Example", false),
+                LintError(8, 5, "$DIKTAT_RULE_SET_ID:kdoc-comments", "${MISSING_KDOC_CLASS_ELEMENTS.warnText()} isValid", false)
         )
     }
 }
