+++ diktat-rules/src/test/kotlin/org/cqfn/diktat/util/FixTestBase.kt
@@ -2,22 +2,24 @@ package org.cqfn.diktat.util
 
 import com.pinterest.ktlint.core.LintError
 import com.pinterest.ktlint.core.Rule
-import com.pinterest.ktlint.core.RuleSet
+import com.pinterest.ktlint.core.RuleSetProvider
 import org.cqfn.diktat.common.config.rules.RulesConfig
 import org.cqfn.diktat.test.framework.processing.TestComparatorUnit
 import org.junit.jupiter.api.Assertions
 
 open class FixTestBase(private val resourceFilePath: String,
-                       private val ruleSupplier: (rulesConfigList: List<RulesConfig>) -> Rule,
-                       rulesConfigList: List<RulesConfig> = emptyList(),
-                       private val cb: (LintError, Boolean) -> Unit = defaultCallback
-) {
-    constructor(resourceFilePath: String, rule: Rule, rulesConfigList: List<RulesConfig>? = emptyList())
-            : this(resourceFilePath, DiktatRuleSetProvider4Test(rule, rulesConfigList).get())
-
+                       private val ruleSetProviderRef: (rulesConfigList: List<RulesConfig>?) -> RuleSetProvider,
+                       private val cb: (LintError, Boolean) -> Unit = defaultCallback,
+                       private val rulesConfigList: List<RulesConfig>? = null) {
     private val testComparatorUnit = TestComparatorUnit(resourceFilePath) { text, fileName ->
-        format(ruleSupplier, text, fileName, rulesConfigList, cb = cb)
+        format(ruleSetProviderRef, text, fileName, rulesConfigList, cb = cb)
     }
+
+    constructor(resourceFilePath: String,
+                ruleSupplier: (rulesConfigList: List<RulesConfig>) -> Rule,
+                rulesConfigList: List<RulesConfig>? = null,
+                cb: (LintError, Boolean) -> Unit = defaultCallback) : this(resourceFilePath,
+            { overrideRulesConfigList -> DiktatRuleSetProvider4Test(ruleSupplier, overrideRulesConfigList) }, cb, rulesConfigList)
 
     protected fun fixAndCompare(expectedPath: String, testPath: String) {
         Assertions.assertTrue(
@@ -30,11 +32,7 @@ open class FixTestBase(private val resourceFilePath: String,
                                 testPath: String,
                                 overrideRulesConfigList: List<RulesConfig>) {
         val testComparatorUnit = TestComparatorUnit(resourceFilePath) { text, fileName ->
-            format(ruleSupplier, text, fileName, overrideRulesConfigList)
-////             fixme: use all rules after https://github.com/cqfn/diKTat/pull/208 is merged
-//            DiktatRuleSetProvider4Test(ruleSet.rules[0], overrideRulesConfigList)
-//                    .get()
-//                    .format(text, fileName)
+            format(ruleSetProviderRef, text, fileName, overrideRulesConfigList)
         }
         Assertions.assertTrue(
                 testComparatorUnit
